multichannel expansion problem solved
the problems happens because of the way how multichannel expansion
happens in patterns: if any of the keys in a pattern receives an array
as value the pattern will instanciate as many Synths as indicated by the
size of the array. Hence, rather than instantiating a multichannel synth
one should take care of the problem by creating a mono version of the
regarding SynthDef beforehand which will automatically be considered
when the sampled pattern is played.

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>