more refactoring
The keyboard sampler needs to know which widget controls a Synth arg.
Hence we collect widget name, arg and the assigned CV in namesCVs.
However, when assigning actions to widgets, it's not feasible to simply
iterate over the Synth's args - we have to iterate over namesCVs in a
forBy loop. That makes sure widgets, argNames and CVs are properly
addressed in the process of assigning actions to widgets.

Signed-off-by: Stefan Nussbaumer <stefan@basislager.org>