bend riddle solved
bend didn't work due to the bend function not being added correctly to
the default bend func. It turned out to be the correct way not to add it
to the MIDIFunc's func by calling addFunc but simply adding it by
calling add on the MIDIFunc with the bendFunc as argument. (lesson
learned: never try to extend a MIDIFunc by calling addFunc on it's
'func' - always call 'add' on the MIDIFunc itself)

Signed-off-by: Stefan Nussbaumer <stefan@basislager.org>